chore: MovieController member, admin, common 으로 분리
